tests/test_operations.py
# -*- coding: utf-8 -*-
import sys
import unittest
import contextlib

from calculator.operations import *
from calculator.exceptions import *
try:
    from io import StringIO
    # python 3
except ImportError:
    from StringIO import StringIO
    # python 2

class TestCalculatorOperations(unittest.TestCase):

    def test_add(self):
        self.assertEqual(add(1), 1)
        self.assertEqual(add(1, 2, 3, 4), 10)
        self.assertEqual(add(1, 2, 3, 4, 5), 15)
        self.assertEqual(add(10, -2), 8)
        self.assertEqual(add(10, 2.5), 12.5)

    def test_subtract(self):
        self.assertEqual(subtract(1), 1)
        self.assertEqual(subtract(10, 1, 3, 3), 3)
        self.assertEqual(subtract(10, 5.5), 4.5)
        self.assertEqual(subtract(10, -2), 12)
        self.assertEqual(subtract(10, 2.5), 7.5)

    def test_multiply(self):
        self.assertEqual(multiply(1), 1)
        self.assertEqual(multiply(2, 3, 5), 30)
        self.assertEqual(multiply(10, 2.5), 25.0)
        self.assertEqual(multiply(10, -2), -20)

    def test_divide(self):
        self.assertEqual(divide(1), 1)
        self.assertEqual(divide(4, 2), 2)
        self.assertEqual(divide(5, 2), 2.5)
        self.assertEqual(divide(10, -2), -5)
        self.assertEqual(divide(12, 2, 3), 2)

    # def test_plot(self):
    #     x = symbols('x')
    #     plot_test = (plot(-x**2, -2, 2))
    #     self.assertEqual(plot_test, 
    #     """
    #     -0.0013 |                       .....  ......                    
    #             |                    ...             ..                  
    #             |                  ..                  ..                
    #             |                ..                      ..              
    #             |              ..                          ..            
    #             |             /                              \           
    #             |            /                                \          
    #             |           /                                  \         
    #     -2.0006 | ---------/------------------------------------\--------
    #             |        ..                                      ..      
    #             |                                                  .     
    #             |       .                                                
    #             |      /                                            .    
    #             |     /                                              \   
    #             |    /                                                \  
    #             |   /                                                  \ 
    #             |  /                                                    \
    #          -4 | /                                                      
    #               -2                     0                          2
    #     """)
    # attempted plot test using textplot and the README example
    # unknown variables: output width, height, etc.
    
    def test_plot_py2(self):
        x = symbols('x')
        saved_stdout = sys.stdout
        try:
            out = StringIO()
            sys.stdout = out
            plot(-x**2, -2, 2) # running the plot function prints out a Textplot, printing to the 'out' variable
            output = out.getvalue().strip()
            assert output == plot(-x**2, -2, 2)
            # again, checking against itself because the values will of course be the same, see comments on py3 test
        finally:
            sys.stdout = saved_stdout
        
    
    def test_plot_py3(self):
        x = symbols('x')
        # assuming symbol provided is 'x'
        temp_stdout = StringIO()
        with contextlib.redirect_stdout(temp_stdout):
            # Python 3 only solution; redirects string output to a temporary variable
            plot(-x**2, -2, 2)
            # plot function by default generates a string output
        output = temp_stdout.getvalue().strip()
        assert output == plot(-x**2, -2, 2)
        # granted, GRANTED, it's checking against itself but it's matching a string to a string
        # Rather than try to type up the whole plot string and be off by like 1 . or /
 